Added basic incrementing, adding items on the fly
Changed the up/down arrows to a single plus symbol. Added AJAXy type
calls to send data to the server and to get a response back. The
prompt auto fills in your most popular choice.

Need to move some of the logic out of the index and into a separate
JavaScript file. Need to create a CSS file as well.